refactor: Rename enums/ to enumdata/
